feat(sql): verify SQL temporal context wiring & fix WAL test coverage
- Refactored `test_append_async_mode` in `src/storage/wal/concurrent_system.rs` to use `wal.shutdown()` for guaranteed flushing, eliminating the race condition and unreachable coverage paths.
- Verified that `convert_sql` correctly extracts temporal clauses and populates `Query.temporal_context`.
- Added `temporal_context()` getter to `Query` struct for better testability.
- Added a unit test for `temporal_context()` in `src/query/builder.rs` to ensure full coverage of the new getter.
- Added `tests/sql_temporal_integration.rs` to verify the end-to-end flow from SQL string to `PhysicalPlan` with temporal context.
- Fixed linting issues.
