--- src/cypher/lexer.rs
@@ -1,4 +1,0 @@
-//! Cypher lexer.
-
-/// Lexer for Cypher queries.
-pub struct Lexer;
