+++ src/index/vector/mod.rs
@@ -43,7 +43,7 @@
 //!
 //! # Examples
 //!
-//! ```rust
+//! ```ignore
 //! use gallifreydb::index::VectorIndex;
 //! use gallifreydb::core::id::NodeId;
 //!
@@ -315,12 +315,10 @@ pub trait VectorIndex: Send + Sync {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```ignore
     /// # use gallifreydb::index::VectorIndex;
     /// # use gallifreydb::core::id::NodeId;
-    /// # use gallifreydb::index::vector::{HnswIndexBuilder, DistanceMetric};
-    /// # fn main() -> gallifreydb::utils::Result<()> {
-    /// # let index = HnswIndexBuilder::new(4, DistanceMetric::Cosine).build()?;
+    /// # fn example(index: &impl VectorIndex) -> gallifreydb::utils::Result<()> {
     /// let node_id = NodeId::new(123).unwrap();
     /// let embedding = vec![0.1, 0.2, 0.3, 0.4];
     /// index.add(node_id, &embedding)?;
@@ -344,12 +342,10 @@ pub trait VectorIndex: Send + Sync {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```ignore
     /// # use gallifreydb::index::VectorIndex;
     /// # use gallifreydb::core::id::NodeId;
-    /// # use gallifreydb::index::vector::{HnswIndexBuilder, DistanceMetric};
-    /// # fn main() -> gallifreydb::utils::Result<()> {
-    /// # let index = HnswIndexBuilder::new(4, DistanceMetric::Cosine).build()?;
+    /// # fn example(index: &impl VectorIndex) -> gallifreydb::utils::Result<()> {
     /// let node_id = NodeId::new(123).unwrap();
     /// index.remove(node_id)?;
     /// # Ok(())
@@ -382,11 +378,9 @@ pub trait VectorIndex: Send + Sync {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```ignore
     /// # use gallifreydb::index::VectorIndex;
-    /// # use gallifreydb::index::vector::{HnswIndexBuilder, DistanceMetric};
-    /// # fn main() -> gallifreydb::utils::Result<()> {
-    /// # let index = HnswIndexBuilder::new(4, DistanceMetric::Cosine).build()?;
+    /// # fn example(index: &impl VectorIndex) -> gallifreydb::utils::Result<()> {
     /// let query = vec![0.5, 0.3, 0.1, 0.9];
     /// let results = index.search(&query, 10)?;
     ///
@@ -432,13 +426,11 @@ pub trait VectorIndex: Send + Sync {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```ignore
     /// # use gallifreydb::index::VectorIndex;
     /// # use gallifreydb::core::id::NodeId;
     /// # use std::collections::HashSet;
-    /// # use gallifreydb::index::vector::{HnswIndexBuilder, DistanceMetric};
-    /// # fn main() -> gallifreydb::utils::Result<()> {
-    /// # let index = HnswIndexBuilder::new(4, DistanceMetric::Cosine).build()?;
+    /// # fn example(index: &impl VectorIndex) -> gallifreydb::utils::Result<()> {
     /// let query = vec![0.5, 0.3, 0.1, 0.9];
     /// let allowed = HashSet::from([NodeId::new(1).unwrap(), NodeId::new(5).unwrap(), NodeId::new(10).unwrap()]);
     ///
@@ -460,13 +452,10 @@ pub trait VectorIndex: Send + Sync {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```ignore
     /// # use gallifreydb::index::VectorIndex;
-    /// # use gallifreydb::index::vector::{HnswIndexBuilder, DistanceMetric};
-    /// # fn main() -> gallifreydb::utils::Result<()> {
-    /// # let index = HnswIndexBuilder::new(4, DistanceMetric::Cosine).build()?;
+    /// # fn example(index: &impl VectorIndex) {
     /// println!("Index contains {} vectors", index.len());
-    /// # Ok(())
     /// # }
     /// ```
     #[must_use]
@@ -478,14 +467,11 @@ pub trait VectorIndex: Send + Sync {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```ignore
     /// # use gallifreydb::index::VectorIndex;
-    /// # use gallifreydb::index::vector::{HnswIndexBuilder, DistanceMetric};
-    /// # fn main() -> gallifreydb::utils::Result<()> {
-    /// # let index = HnswIndexBuilder::new(4, DistanceMetric::Cosine).build()?;
+    /// # fn example(index: &impl VectorIndex) {
     /// let dims = index.dimensions();
     /// println!("This index accepts {}-dimensional vectors", dims);
-    /// # Ok(())
     /// # }
     /// ```
     #[must_use]
@@ -508,11 +494,9 @@ pub trait VectorIndex: Send + Sync {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```ignore
     /// # use gallifreydb::index::{VectorIndex, DistanceMetric};
-    /// # use gallifreydb::index::vector::HnswIndexBuilder;
-    /// # fn main() -> gallifreydb::utils::Result<()> {
-    /// # let index = HnswIndexBuilder::new(4, DistanceMetric::Cosine).build()?;
+    /// # fn example(index: &impl VectorIndex) {
     /// match index.distance_metric() {
     ///     DistanceMetric::Cosine => println!("Using cosine similarity"),
     ///     DistanceMetric::Euclidean => println!("Using Euclidean distance"),
@@ -521,7 +505,6 @@ pub trait VectorIndex: Send + Sync {
     ///     DistanceMetric::Hamming => println!("Using Hamming distance"),
     ///     DistanceMetric::Tanimoto => println!("Using Tanimoto similarity"),
     /// }
-    /// # Ok(())
     /// # }
     /// ```
     #[must_use]
@@ -533,15 +516,12 @@ pub trait VectorIndex: Send + Sync {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```ignore
     /// # use gallifreydb::index::VectorIndex;
-    /// # use gallifreydb::index::vector::{HnswIndexBuilder, DistanceMetric};
-    /// # fn main() -> gallifreydb::utils::Result<()> {
-    /// # let index = HnswIndexBuilder::new(4, DistanceMetric::Cosine).build()?;
+    /// # fn example(index: &impl VectorIndex) {
     /// if index.is_empty() {
     ///     println!("No vectors indexed yet");
     /// }
-    /// # Ok(())
     /// # }
     /// ```
     #[must_use]
