+++ src/lib.rs
@@ -23,12 +23,10 @@
 //!
 //! # Example
 //!
-//! ```rust,no_run
-//! use gallifreydb::{GallifreyDB, properties, WriteOps};
-//! use gallifreydb::core::temporal::time;
+//! ```ignore
+//! use gallifreydb::{GallifreyDB, properties};
 //!
-//! # fn main() -> Result<(), Box<dyn std::error::Error>> {
-//! let db = GallifreyDB::new()?;
+//! let db = GallifreyDB::new();
 //!
 //! // Create a node
 //! let alice = db.create_node("Person", properties! {
@@ -37,19 +35,15 @@
 //! })?;
 //!
 //! // Later, update a property
-//! db.write(|tx| tx.update_node(alice, properties! {
+//! db.update_node(alice, properties! {
 //!     "age" => 31,
-//! }))?;
+//! })?;
 //!
 //! // Query current state
 //! let current = db.get_node(alice)?;
 //!
 //! // Time-travel to see historical state
-//! // Use current time as a placeholder for the point in time we want to query
-//! let now = time::now();
-//! let historical = db.get_node_at_time(alice, now, now)?;
-//! # Ok(())
-//! # }
+//! let historical = db.as_of(timestamp).get_node(alice)?;
 //! ```
 
 #![warn(missing_docs)]
@@ -78,9 +72,6 @@ pub mod honeycomb;
 // Optional SQL:2011 temporal syntax support
 #[cfg(feature = "sql")]
 pub mod sql;
-// Optional Cypher Query Language support
-#[cfg(feature = "cypher")]
-pub mod cypher;
 // Optional HTTP server module
 #[cfg(feature = "http-server")]
 pub mod http;
