+++ src/query/builder.rs
@@ -6,19 +6,7 @@
 //!
 //! # Example
 //!
-//! ```rust
-//! use gallifreydb::query::QueryBuilder;
-//! use gallifreydb::core::NodeId;
-//! use gallifreydb::core::temporal::Timestamp;
-//!
-//! # fn main() -> Result<(), Box<dyn std::error::Error>> {
-//! let alice_id = NodeId::new(1)?;
-//! let bob_id = NodeId::new(2)?;
-//! let node_id = NodeId::new(3)?;
-//! let embedding = vec![0.1, 0.2, 0.3];
-//! let valid_time = Timestamp::new(1000, 0)?;
-//! let tx_time = Timestamp::new(2000, 0)?;
-//!
+//! ```rust,ignore
 //! // Build a graph + vector query
 //! let query = QueryBuilder::new()
 //!     .start(alice_id)
@@ -46,8 +34,6 @@
 //!         .label_filter("Person")
 //!         .finish()
 //!     .build();
-//! # Ok(())
-//! # }
 //! ```
 
 use std::marker::PhantomData;
@@ -204,11 +190,7 @@ impl QueryBuilder<state::Initial> {
     ///
     /// # Example
     ///
-    /// ```rust
-    /// use gallifreydb::query::QueryBuilder;
-    /// use gallifreydb::index::vector::DistanceMetric;
-    ///
-    /// let embedding = vec![0.1, 0.2, 0.3];
+    /// ```ignore
     /// let query = QueryBuilder::new()
     ///     .find_similar_builder(&embedding, 10)
     ///         .property("title_embedding")
@@ -313,14 +295,7 @@ impl QueryBuilder<state::HasNodes> {
     ///
     /// # Example
     ///
-    /// ```rust
-    /// use gallifreydb::query::QueryBuilder;
-    /// use gallifreydb::core::NodeId;
-    ///
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// let alice_id = NodeId::new(1)?;
-    /// let embedding = vec![0.1, 0.2, 0.3];
-    ///
+    /// ```ignore
     /// let query = QueryBuilder::new()
     ///     .start(alice_id)
     ///     .traverse("KNOWS")
@@ -328,8 +303,6 @@ impl QueryBuilder<state::HasNodes> {
     ///         .property("custom_embedding")
     ///         .finish()
     ///     .build();
-    /// # Ok(())
-    /// # }
     /// ```
     pub fn rank_by_similarity_builder(
         self,
@@ -367,19 +340,11 @@ impl QueryBuilder<state::HasNodes> {
     ///
     /// # Example
     ///
-    /// ```rust,no_run
-    /// # use gallifreydb::GallifreyDB;
-    /// # use gallifreydb::core::NodeId;
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let db = GallifreyDB::new()?;
-    /// # let node_id = NodeId::new(1)?;
-    /// # let bob_id = NodeId::new(2)?;
+    /// ```ignore
     /// let results = db.query()
     ///     .start(node_id)
     ///     .similar_to(bob_id, 10)
-    ///     .execute(&db)?;
-    /// # Ok(())
-    /// # }
+    ///     .execute()?;
     /// ```
     ///
     /// # See Also
@@ -416,13 +381,7 @@ impl QueryBuilder<state::HasNodes> {
     ///
     /// # Example
     ///
-    /// ```rust,no_run
-    /// # use gallifreydb::GallifreyDB;
-    /// # use gallifreydb::core::NodeId;
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let db = GallifreyDB::new()?;
-    /// # let alice_id = NodeId::new(1)?;
-    /// # let bob_id = NodeId::new(2)?;
+    /// ```ignore
     /// // Find similar documents using custom embedding
     /// let results = db.query()
     ///     .start(alice_id)
@@ -430,9 +389,7 @@ impl QueryBuilder<state::HasNodes> {
     ///         .property("custom_embedding")
     ///         .label_filter("Document")
     ///         .finish()
-    ///     .execute(&db)?;
-    /// # Ok(())
-    /// # }
+    ///     .execute()?;
     /// ```
     ///
     /// # Panics
@@ -605,22 +562,13 @@ impl<S: QueryState> QueryBuilder<S> {
     ///
     /// # Example
     ///
-    /// ```rust
-    /// # use gallifreydb::query::QueryBuilder;
-    /// # use gallifreydb::core::NodeId;
-    /// # use gallifreydb::core::temporal::Timestamp;
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let timestamp = Timestamp::new(1000, 0)?;
-    /// # let tx_time = Timestamp::new(2000, 0)?;
-    /// # let node_id = NodeId::new(1)?;
-    /// let query = QueryBuilder::new()
+    /// ```ignore
+    /// let results = db.query()
     ///     .as_of(timestamp, tx_time)
     ///     .start(node_id)
     ///     .traverse("KNOWS")
     ///     .with_provenance()
     ///     .build();
-    /// # Ok(())
-    /// # }
     /// ```
     #[must_use]
     pub fn with_provenance(mut self) -> Self {
@@ -651,14 +599,10 @@ impl<S: QueryState> QueryBuilder<S> {
     ///
     /// # Example
     ///
-    /// ```rust,no_run
+    /// ```ignore
     /// use gallifreydb::GallifreyDB;
-    /// use gallifreydb::core::NodeId;
     ///
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// let db = GallifreyDB::new()?;
-    /// let alice_id = NodeId::new(1)?;
-    ///
+    /// let db = GallifreyDB::new();
     /// let results = db.query()
     ///     .start(alice_id)
     ///     .traverse("KNOWS")
@@ -668,8 +612,6 @@ impl<S: QueryState> QueryBuilder<S> {
     ///     let row = row?;
     ///     println!("Found: {:?}", row.entity);
     /// }
-    /// # Ok(())
-    /// # }
     /// ```
     pub fn execute(
         self,
@@ -712,12 +654,7 @@ impl<S: QueryState> QueryBuilder<S> {
 ///
 /// # Example
 ///
-/// ```rust
-/// # use gallifreydb::query::QueryBuilder;
-/// # use gallifreydb::core::NodeId;
-/// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// # let alice_id = NodeId::new(1)?;
-/// # let bob_id = NodeId::new(2)?;
+/// ```ignore
 /// let query = QueryBuilder::new()
 ///     .start(alice_id)
 ///     .similar_to_builder(bob_id, 10)
@@ -725,8 +662,6 @@ impl<S: QueryState> QueryBuilder<S> {
 ///         .label_filter("Person")
 ///         .finish()
 ///     .build();
-/// # Ok(())
-/// # }
 /// ```
 #[must_use = "builders do nothing unless you call finish()"]
 pub struct SimilarToBuilder<S: QueryState> {
@@ -757,18 +692,10 @@ impl<S: QueryState> SimilarToBuilder<S> {
     ///
     /// # Example
     ///
-    /// ```rust
-    /// # use gallifreydb::query::QueryBuilder;
-    /// # use gallifreydb::core::NodeId;
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let node_id = NodeId::new(1)?;
-    /// QueryBuilder::new()
-    ///     .start(node_id)
-    ///     .similar_to_builder(node_id, 10)
+    /// ```ignore
+    /// .similar_to_builder(node_id, 10)
     ///     .property("custom_embedding")
-    ///     .finish();
-    /// # Ok(())
-    /// # }
+    ///     .finish()
     /// ```
     pub fn property(mut self, key: impl Into<String>) -> Self {
         self.property_key = Some(key.into());
@@ -779,18 +706,10 @@ impl<S: QueryState> SimilarToBuilder<S> {
     ///
     /// # Example
     ///
-    /// ```rust
-    /// # use gallifreydb::query::QueryBuilder;
-    /// # use gallifreydb::core::NodeId;
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let node_id = NodeId::new(1)?;
-    /// QueryBuilder::new()
-    ///     .start(node_id)
-    ///     .similar_to_builder(node_id, 10)
+    /// ```ignore
+    /// .similar_to_builder(node_id, 10)
     ///     .label_filter("Document")
-    ///     .finish();
-    /// # Ok(())
-    /// # }
+    ///     .finish()
     /// ```
     pub fn label_filter(mut self, label: impl Into<String>) -> Self {
         self.label_filter = Some(label.into());
@@ -814,12 +733,7 @@ impl<S: QueryState> SimilarToBuilder<S> {
 ///
 /// # Example
 ///
-/// ```rust
-/// # use gallifreydb::query::QueryBuilder;
-/// # use gallifreydb::core::NodeId;
-/// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// # let alice_id = NodeId::new(1)?;
-/// # let embedding = vec![0.1, 0.2, 0.3];
+/// ```ignore
 /// let query = QueryBuilder::new()
 ///     .start(alice_id)
 ///     .traverse("KNOWS")
@@ -827,8 +741,6 @@ impl<S: QueryState> SimilarToBuilder<S> {
 ///         .property("custom_embedding")
 ///         .finish()
 ///     .build();
-/// # Ok(())
-/// # }
 /// ```
 #[must_use = "builders do nothing unless you call finish()"]
 pub struct RankBySimilarityBuilder<S: QueryState> {
@@ -854,20 +766,10 @@ impl<S: QueryState> RankBySimilarityBuilder<S> {
     ///
     /// # Example
     ///
-    /// ```rust
-    /// # use gallifreydb::query::QueryBuilder;
-    /// # use gallifreydb::core::NodeId;
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let alice_id = NodeId::new(1)?;
-    /// # let embedding = vec![0.1, 0.2, 0.3];
-    /// QueryBuilder::new()
-    ///     .start(alice_id)
-    ///     .traverse("KNOWS")
-    ///     .rank_by_similarity_builder(&embedding, 10)
+    /// ```ignore
+    /// .rank_by_similarity_builder(&embedding, 10)
     ///     .property("custom_embedding")
-    ///     .finish();
-    /// # Ok(())
-    /// # }
+    ///     .finish()
     /// ```
     pub fn property(mut self, key: impl Into<String>) -> Self {
         self.property_key = Some(key.into());
@@ -890,18 +792,13 @@ impl<S: QueryState> RankBySimilarityBuilder<S> {
 ///
 /// # Example
 ///
-/// ```rust
-/// # use gallifreydb::query::QueryBuilder;
-/// # use gallifreydb::index::vector::DistanceMetric;
-/// # fn main() {
-/// # let embedding = vec![0.1, 0.2, 0.3];
+/// ```ignore
 /// let query = QueryBuilder::new()
 ///     .find_similar_builder(&embedding, 10)
 ///         .property("title_embedding")
 ///         .metric(DistanceMetric::Euclidean)
 ///         .finish()
 ///     .build();
-/// # }
 /// ```
 #[must_use = "builders do nothing unless you call finish()"]
 pub struct FindSimilarBuilder {
@@ -929,15 +826,10 @@ impl FindSimilarBuilder {
     ///
     /// # Example
     ///
-    /// ```rust
-    /// # use gallifreydb::query::QueryBuilder;
-    /// # fn main() {
-    /// # let embedding = vec![0.1, 0.2, 0.3];
-    /// QueryBuilder::new()
-    ///     .find_similar_builder(&embedding, 10)
+    /// ```ignore
+    /// .find_similar_builder(&embedding, 10)
     ///     .property("body_embedding")
-    ///     .finish();
-    /// # }
+    ///     .finish()
     /// ```
     pub fn property(mut self, key: impl Into<String>) -> Self {
         self.property_key = Some(key.into());
@@ -950,16 +842,10 @@ impl FindSimilarBuilder {
     ///
     /// # Example
     ///
-    /// ```rust
-    /// # use gallifreydb::query::QueryBuilder;
-    /// # use gallifreydb::index::vector::DistanceMetric;
-    /// # fn main() {
-    /// # let embedding = vec![0.1, 0.2, 0.3];
-    /// QueryBuilder::new()
-    ///     .find_similar_builder(&embedding, 10)
+    /// ```ignore
+    /// .find_similar_builder(&embedding, 10)
     ///     .metric(DistanceMetric::Euclidean)
-    ///     .finish();
-    /// # }
+    ///     .finish()
     /// ```
     pub fn metric(mut self, metric: DistanceMetric) -> Self {
         self.metric = metric;
@@ -1391,5 +1277,21 @@ mod tests {
             }
             _ => panic!("Expected VectorSearch operation"),
         }
+    }
+
+    #[test]
+    fn test_temporal_context_getter() {
+        let query = QueryBuilder::new()
+            .as_of(1000.into(), 2000.into())
+            .start(test_node_id())
+            .build();
+
+        assert!(query.is_temporal());
+        assert!(query.temporal_context().is_some());
+
+        let ctx = query.temporal_context().unwrap();
+        assert!(ctx.as_of.is_some());
+        assert_eq!(ctx.as_of.unwrap().0.wallclock(), 1000);
+        assert_eq!(ctx.as_of.unwrap().1.wallclock(), 2000);
     }
 }
