--- src/storage/wal/serialization.rs
@@ -1,483 +1,0 @@
-//! Serialization logic for WAL entries.
-
-use super::entry::{WalEntry, WalOperation};
-use crate::core::interning::InternedString;
-use crate::utils::error::Result;
-
-/// Helper to serialize an InternedString into the buffer (4-byte ID)
-#[inline(always)]
-fn serialize_interned_string(s: InternedString, buffer: &mut Vec<u8>) {
-    buffer.extend_from_slice(&s.as_u32().to_le_bytes());
-}
-
-/// Estimate the required buffer capacity for serializing a WAL entry.
-///
-/// This function provides an upper-bound estimate of the buffer size needed
-/// to serialize a WAL entry, allowing pre-allocation to avoid reallocations
-/// during the hot-path serialization process.
-///
-/// # Size Breakdown
-///
-/// Fixed overhead (per entry):
-/// - LSN: 8 bytes
-/// - Timestamp (HybridTimestamp): 12 bytes
-/// - Checksum: 4 bytes
-/// - Total fixed: 24 bytes
-///
-/// Variable sizes by operation:
-/// - `CreateNode`: 1 (op type) + 8 (node_id) + 4 (label len) + label bytes +
-///   properties size + 48 (BiTemporalInterval)
-/// - `CreateEdge`: 1 (op type) + 8 (edge_id) + 8 (source) + 8 (target) +
-///   4 (label len) + label bytes + properties size + 48 (BiTemporalInterval)
-/// - `UpdateNode`: 1 (op type) + 8 (node_id) + 8 (version_id) + 4 (label len) +
-///   label bytes + properties size + 48 (BiTemporalInterval)
-/// - `UpdateEdge`: 1 (op type) + 8 (edge_id) + 8 (version_id) + 4 (label len) +
-///   label bytes + properties size + 48 (BiTemporalInterval)
-/// - `DeleteNode`: 1 (op type) + 8 (node_id) + 48 (BiTemporalInterval) = 57 bytes
-/// - `DeleteEdge`: 1 (op type) + 8 (edge_id) + 48 (BiTemporalInterval) = 57 bytes
-/// - `Checkpoint`: 1 (op type) + 8 (lsn) + 12 (timestamp) = 21 bytes
-///
-/// # Returns
-///
-/// An estimated capacity in bytes. The estimate is conservative (may slightly
-/// over-allocate) to ensure the buffer doesn't need to grow during serialization.
-///
-/// # Performance Impact
-///
-/// Pre-allocating the correct capacity eliminates dynamic reallocation overhead,
-/// which is especially important for the high-throughput WAL write path. Typical
-/// savings are 10-30% reduction in allocation overhead for property-heavy operations.
-pub(crate) fn estimate_entry_capacity(operation: &WalOperation) -> usize {
-    // Fixed overhead: LSN (8) + Timestamp (12) + Checksum (4)
-    const FIXED_OVERHEAD: usize = 24;
-    // BiTemporalInterval is always 48 bytes (2 * TimeRange, each 24 bytes)
-    const TEMPORAL_SIZE: usize = 48;
-
-    let variable_size = match operation {
-        WalOperation::CreateNode { properties, .. } => {
-            // op type (1) + node_id (8) + label (4-byte InternedString ID) + properties + temporal (48)
-            let base = 1 + 8 + 4 + TEMPORAL_SIZE;
-            base + properties.serialized_size()
-        }
-        WalOperation::CreateEdge { properties, .. } => {
-            // op type (1) + edge_id (8) + source (8) + target (8) + label (4-byte InternedString ID) + properties + temporal (48)
-            let base = 1 + 8 + 8 + 8 + 4 + TEMPORAL_SIZE;
-            base + properties.serialized_size()
-        }
-        WalOperation::UpdateNode { properties, .. } => {
-            // op type (1) + node_id (8) + version_id (8) + label (4-byte InternedString ID) + properties + temporal (48)
-            let base = 1 + 8 + 8 + 4 + TEMPORAL_SIZE;
-            base + properties.serialized_size()
-        }
-        WalOperation::UpdateEdge { properties, .. } => {
-            // op type (1) + edge_id (8) + version_id (8) + label (4-byte InternedString ID) + properties + temporal (48)
-            let base = 1 + 8 + 8 + 4 + TEMPORAL_SIZE;
-            base + properties.serialized_size()
-        }
-        WalOperation::DeleteNode { .. } => {
-            // op type (1) + node_id (8) + temporal (48)
-            1 + 8 + TEMPORAL_SIZE
-        }
-        WalOperation::DeleteEdge { .. } => {
-            // op type (1) + edge_id (8) + temporal (48)
-            1 + 8 + TEMPORAL_SIZE
-        }
-        WalOperation::Checkpoint { .. } => {
-            // op type (1) + lsn (8) + timestamp (12)
-            1 + 8 + 12
-        }
-    };
-
-    FIXED_OVERHEAD + variable_size
-}
-
-/// Serialize a WAL entry with CRC32 checksum into the provided buffer
-///
-/// This function reuses the provided buffer to avoid per-entry allocation.
-/// The caller should clear the buffer before calling this function to maintain its capacity.
-pub(crate) fn serialize_entry_into(entry: &WalEntry, buffer: &mut Vec<u8>) -> Result<()> {
-    // Write LSN (8 bytes)
-    buffer.extend_from_slice(&entry.lsn.0.to_le_bytes());
-
-    // Write timestamp (12 bytes: Phase 2 HybridTimestamp)
-    entry.timestamp.serialize_into(buffer);
-
-    // Reserve space for checksum (4 bytes) - will fill in later
-    let checksum_offset = buffer.len();
-    buffer.extend_from_slice(&[0u8; 4]);
-
-    // Write operation type and data with full serialization
-    match &entry.operation {
-        WalOperation::CreateNode {
-            node_id,
-            label,
-            properties,
-            temporal,
-        } => {
-            buffer.push(1); // operation type
-            buffer.extend_from_slice(&node_id.as_u64().to_le_bytes());
-            serialize_interned_string(*label, buffer);
-            properties.serialize_into(buffer)?;
-            temporal.serialize_into(buffer);
-        }
-        WalOperation::CreateEdge {
-            edge_id,
-            source,
-            target,
-            label,
-            properties,
-            temporal,
-        } => {
-            buffer.push(2); // operation type
-            buffer.extend_from_slice(&edge_id.as_u64().to_le_bytes());
-            buffer.extend_from_slice(&source.as_u64().to_le_bytes());
-            buffer.extend_from_slice(&target.as_u64().to_le_bytes());
-            serialize_interned_string(*label, buffer);
-            properties.serialize_into(buffer)?;
-            temporal.serialize_into(buffer);
-        }
-        WalOperation::UpdateNode {
-            node_id,
-            version_id,
-            label,
-            properties,
-            temporal,
-        } => {
-            buffer.push(3); // operation type
-            buffer.extend_from_slice(&node_id.as_u64().to_le_bytes());
-            buffer.extend_from_slice(&version_id.as_u64().to_le_bytes());
-            serialize_interned_string(*label, buffer);
-            properties.serialize_into(buffer)?;
-            temporal.serialize_into(buffer);
-        }
-        WalOperation::UpdateEdge {
-            edge_id,
-            version_id,
-            label,
-            properties,
-            temporal,
-        } => {
-            buffer.push(4); // operation type
-            buffer.extend_from_slice(&edge_id.as_u64().to_le_bytes());
-            buffer.extend_from_slice(&version_id.as_u64().to_le_bytes());
-            serialize_interned_string(*label, buffer);
-            properties.serialize_into(buffer)?;
-            temporal.serialize_into(buffer);
-        }
-        WalOperation::DeleteNode { node_id, temporal } => {
-            buffer.push(6); // operation type
-            buffer.extend_from_slice(&node_id.as_u64().to_le_bytes());
-            temporal.serialize_into(buffer);
-        }
-        WalOperation::DeleteEdge { edge_id, temporal } => {
-            buffer.push(7); // operation type
-            buffer.extend_from_slice(&edge_id.as_u64().to_le_bytes());
-            temporal.serialize_into(buffer);
-        }
-        WalOperation::Checkpoint { lsn, timestamp } => {
-            buffer.push(5); // operation type
-            buffer.extend_from_slice(&lsn.0.to_le_bytes());
-            // Phase 2: Use HybridTimestamp serialization
-            timestamp.serialize_into(buffer);
-        }
-    }
-
-    // Compute CRC32 over everything except the checksum field
-    let mut hasher = crc32fast::Hasher::new();
-    hasher.update(&buffer[0..checksum_offset]); // LSN + timestamp
-    hasher.update(&buffer[checksum_offset + 4..]); // Operation data
-    let checksum = hasher.finalize();
-
-    // Write the checksum into the reserved space
-    buffer[checksum_offset..checksum_offset + 4].copy_from_slice(&checksum.to_le_bytes());
-
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::core::EdgeId;
-    use crate::core::NodeId;
-    use crate::core::interning::GLOBAL_INTERNER;
-    use crate::core::property::PropertyMapBuilder;
-    use crate::core::temporal::{BiTemporalInterval, Timestamp};
-    use crate::storage::wal::entry::LSN; // Imported only for tests
-
-    /// Helper to create a test temporal interval
-    fn test_temporal() -> BiTemporalInterval {
-        use crate::core::hlc::HybridTimestamp;
-        let timestamp = HybridTimestamp::new_unchecked(1000000, 0);
-        BiTemporalInterval::current(timestamp)
-    }
-
-    /// Helper to create a test timestamp
-    fn test_timestamp() -> Timestamp {
-        use crate::core::hlc::HybridTimestamp;
-        HybridTimestamp::new_unchecked(1000000, 0)
-    }
-
-    #[test]
-    fn test_estimate_capacity_checkpoint() {
-        // Checkpoint: op type (1) + lsn (8) + timestamp (12) = 21 bytes
-        // Fixed overhead: LSN (8) + Timestamp (12) + Checksum (4) = 24 bytes
-        // Total: 45 bytes
-        let op = WalOperation::Checkpoint {
-            lsn: LSN(1),
-            timestamp: test_timestamp(),
-        };
-
-        let estimated = estimate_entry_capacity(&op);
-        assert_eq!(estimated, 45, "Checkpoint should be exactly 45 bytes");
-
-        // Verify by actually serializing
-        let entry = WalEntry::new(LSN(1), op);
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-        assert!(
-            buffer.len() <= estimated,
-            "Actual size {} should not exceed estimate {}",
-            buffer.len(),
-            estimated
-        );
-    }
-
-    #[test]
-    fn test_estimate_capacity_delete_node() {
-        // DeleteNode: op type (1) + node_id (8) + temporal (48) = 57 bytes
-        // Fixed overhead: 24 bytes
-        // Total: 81 bytes
-        let op = WalOperation::DeleteNode {
-            node_id: NodeId::new(1).unwrap(),
-            temporal: test_temporal(),
-        };
-
-        let estimated = estimate_entry_capacity(&op);
-        assert_eq!(estimated, 81, "DeleteNode should be exactly 81 bytes");
-
-        // Verify by actually serializing
-        let entry = WalEntry::new(LSN(1), op);
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-        assert!(
-            buffer.len() <= estimated,
-            "Actual size {} should not exceed estimate {}",
-            buffer.len(),
-            estimated
-        );
-    }
-
-    #[test]
-    fn test_estimate_capacity_delete_edge() {
-        // DeleteEdge: op type (1) + edge_id (8) + temporal (48) = 57 bytes
-        // Fixed overhead: 24 bytes
-        // Total: 81 bytes
-        let op = WalOperation::DeleteEdge {
-            edge_id: EdgeId::new(1).unwrap(),
-            temporal: test_temporal(),
-        };
-
-        let estimated = estimate_entry_capacity(&op);
-        assert_eq!(estimated, 81, "DeleteEdge should be exactly 81 bytes");
-
-        // Verify by actually serializing
-        let entry = WalEntry::new(LSN(1), op);
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-        assert!(
-            buffer.len() <= estimated,
-            "Actual size {} should not exceed estimate {}",
-            buffer.len(),
-            estimated
-        );
-    }
-
-    #[test]
-    fn test_estimate_capacity_create_node_empty_properties() {
-        // CreateNode with empty properties:
-        // Fixed: 24 bytes (LSN + Timestamp + Checksum)
-        // op type (1) + node_id (8) + label (4-byte InternedString) + properties (4 for empty count) + temporal (48)
-        // = 24 + 1 + 8 + 4 + 4 + 48 = 89 bytes
-        let op = WalOperation::CreateNode {
-            node_id: NodeId::new(1).unwrap(),
-            label: GLOBAL_INTERNER.intern("test").unwrap(),
-            properties: PropertyMapBuilder::new().build(),
-            temporal: test_temporal(),
-        };
-
-        let estimated = estimate_entry_capacity(&op);
-        assert_eq!(
-            estimated, 89,
-            "CreateNode with empty properties should be 89 bytes"
-        );
-
-        // Verify by actually serializing
-        let entry = WalEntry::new(LSN(1), op);
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-        assert!(
-            buffer.len() <= estimated,
-            "Actual size {} should not exceed estimate {}",
-            buffer.len(),
-            estimated
-        );
-    }
-
-    #[test]
-    fn test_estimate_capacity_create_node_with_properties() {
-        // CreateNode with properties
-        let properties = PropertyMapBuilder::new()
-            .insert("name", "Alice")
-            .insert("age", 30)
-            .insert("score", 95.5)
-            .build();
-
-        let op = WalOperation::CreateNode {
-            node_id: NodeId::new(1).unwrap(),
-            label: GLOBAL_INTERNER.intern("Person").unwrap(),
-            properties,
-            temporal: test_temporal(),
-        };
-
-        let estimated = estimate_entry_capacity(&op);
-
-        // Verify by actually serializing
-        let entry = WalEntry::new(LSN(1), op);
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        assert!(
-            buffer.len() <= estimated,
-            "Actual size {} should not exceed estimate {}",
-            buffer.len(),
-            estimated
-        );
-
-        // The estimate should be reasonably close (not wildly over-allocated)
-        let overhead_ratio = estimated as f64 / buffer.len() as f64;
-        assert!(
-            overhead_ratio <= 1.5,
-            "Estimate {} should not be more than 50% over actual size {}",
-            estimated,
-            buffer.len()
-        );
-    }
-
-    #[test]
-    fn test_estimate_capacity_create_edge() {
-        // CreateEdge with properties
-        let properties = PropertyMapBuilder::new()
-            .insert("weight", 1.5)
-            .insert("type", "FRIEND")
-            .build();
-
-        let op = WalOperation::CreateEdge {
-            edge_id: EdgeId::new(1).unwrap(),
-            source: NodeId::new(1).unwrap(),
-            target: NodeId::new(2).unwrap(),
-            label: GLOBAL_INTERNER.intern("KNOWS").unwrap(),
-            properties,
-            temporal: test_temporal(),
-        };
-
-        let estimated = estimate_entry_capacity(&op);
-
-        // Verify by actually serializing
-        let entry = WalEntry::new(LSN(1), op);
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        assert!(
-            buffer.len() <= estimated,
-            "Actual size {} should not exceed estimate {}",
-            buffer.len(),
-            estimated
-        );
-
-        let overhead_ratio = estimated as f64 / buffer.len() as f64;
-        assert!(
-            overhead_ratio <= 1.5,
-            "Estimate {} should not be more than 50% over actual size {}",
-            estimated,
-            buffer.len()
-        );
-    }
-
-    #[test]
-    fn test_estimate_capacity_with_vector_property() {
-        // CreateNode with vector property
-        let embedding = vec![0.1, 0.2, 0.3, 0.4];
-        let properties = PropertyMapBuilder::new()
-            .insert_vector("embedding", &embedding)
-            .build();
-
-        let op = WalOperation::CreateNode {
-            node_id: NodeId::new(1).unwrap(),
-            label: GLOBAL_INTERNER.intern("Document").unwrap(),
-            properties,
-            temporal: test_temporal(),
-        };
-
-        let estimated = estimate_entry_capacity(&op);
-
-        // Verify by actually serializing
-        let entry = WalEntry::new(LSN(1), op);
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        assert!(
-            buffer.len() <= estimated,
-            "Actual size {} should not exceed estimate {}",
-            buffer.len(),
-            estimated
-        );
-
-        let overhead_ratio = estimated as f64 / buffer.len() as f64;
-        assert!(
-            overhead_ratio <= 1.5,
-            "Estimate {} should not be more than 50% over actual size {}",
-            estimated,
-            buffer.len()
-        );
-    }
-
-    #[test]
-    fn test_estimate_capacity_large_properties() {
-        // Test with large property map to ensure estimate handles it
-        let mut builder = PropertyMapBuilder::new();
-        for i in 0..50 {
-            builder = builder.insert(&format!("key_{}", i), i);
-        }
-        let properties = builder.build();
-
-        let op = WalOperation::UpdateNode {
-            node_id: NodeId::new(1).unwrap(),
-            version_id: crate::core::VersionId::new(1).unwrap(),
-            label: GLOBAL_INTERNER.intern("LargeNode").unwrap(),
-            properties,
-            temporal: test_temporal(),
-        };
-
-        let estimated = estimate_entry_capacity(&op);
-
-        // Verify by actually serializing
-        let entry = WalEntry::new(LSN(1), op);
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        assert!(
-            buffer.len() <= estimated,
-            "Actual size {} should not exceed estimate {}",
-            buffer.len(),
-            estimated
-        );
-
-        let overhead_ratio = estimated as f64 / buffer.len() as f64;
-        assert!(
-            overhead_ratio <= 1.5,
-            "Estimate {} should not be more than 50% over actual size {}",
-            estimated,
-            buffer.len()
-        );
-    }
-}
